+++ src/Phan/Language/FileRef.php
@@ -189,7 +189,7 @@ class FileRef implements \Serializable
 
     public function serialize(): string
     {
-        return $this->__toString();
+        return \serialize($this->__serialize());
     }
 
     /**
@@ -197,10 +197,23 @@ class FileRef implements \Serializable
      */
     public function unserialize($serialized): void
     {
-        $map = \explode(':', $serialized);
-        $this->file = $map[0];
-        $this->line_number_start = (int)$map[1];
-        $this->line_number_end = (int)($map[2] ?? 0);
+        $this->__unserialize(\unserialize($serialized));
+    }
+
+    /**
+     * @return array{0:string, 1:int, 2:int}
+     */
+    public function __serialize(): array
+    {
+        return [$this->file, $this->line_number_start, $this->line_number_end];
+    }
+
+    /**
+     * @param array{0:string, 1:int, 2:int} $data
+     */
+    public function __unserialize(array $data): void
+    {
+        [$this->file, $this->line_number_start, $this->line_number_end] = $data;
     }
 
     /**